[YE-0] Fix return type.

+++ src/main/java/be/yildizgames/module/vfs/VfsFile.java
@@ -32,6 +32,6 @@ package be.yildizgames.module.vfs;
  */
 public interface VfsFile {
 
-    long getSize();
+    int getSize();
 
 }
